Server/FloraSeven_AI/requirements.txt
tensorflow==2.15.0
numpy>=1.22.0
pillow>=9.0.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
tqdm>=4.62.0
kaggle>=1.5.12
requests>=2.27.0
ipykernel>=6.0.0
notebook>=6.4.0
tensorboard>=2.8.0
